requirements.txt
qdrant-client
cohere
python-dotenv
ollama
requests
fastapi
uvicorn
ragas
openai